fix(sqlite-writer): cast len to int64 before multiply to prevent overflow
text_serial_type computed len*PAIR_LEN in 32-bit int before widening to
int64_t, so a TEXT length >= INT_MAX/2+1 overflowed (UB) into a negative,
sign-extended serial type -> put_varint emitting 10 bytes into the 9-byte
vbuf in rec_add_text (CWE-787, latent behind the 100 MB file cap). Cast len
to int64_t before the multiply, mirroring the correct sibling rec_add_blob.

+++ internal/cbm/sqlite_writer.c
@@ -209,7 +209,7 @@ static int varint_len(int64_t value) {
 
 // SQLite serial type for a TEXT value
 static int64_t text_serial_type(int len) {
-    return (len * PAIR_LEN) + TEXT_SERIAL_BASE;
+    return ((int64_t)len * PAIR_LEN) + TEXT_SERIAL_BASE;
 }
 
 // Test seam: exposes the file-static text_serial_type so its 64-bit numeric
